feat(browser): Do not include metrics in base CDN bundle
Metrics are only included when performance is included,
reducing the base bundle size.

We always expose a shim so there is no API breakage, it just does nothing.
