+++ packages/browser/src/index.bundle.feedback.ts
@@ -1,4 +1,4 @@
-import { browserTracingIntegrationShim, replayIntegrationShim } from '@sentry-internal/integration-shims';
+import { browserTracingIntegrationShim, metricsShim, replayIntegrationShim } from '@sentry-internal/integration-shims';
 import { feedbackAsyncIntegration } from './feedbackAsync';
 
 export * from './index.bundle.base';
@@ -10,6 +10,7 @@ export {
   feedbackAsyncIntegration as feedbackAsyncIntegration,
   feedbackAsyncIntegration as feedbackIntegration,
   replayIntegrationShim as replayIntegration,
+  metricsShim as metrics,
 };
 
 export { captureFeedback } from '@sentry/core';
