+++ packages/browser/src/index.bundle.tracing.replay.feedback.ts
@@ -4,6 +4,8 @@ registerSpanErrorInstrumentation();
 
 export * from './index.bundle.base';
 
+export * from './metrics';
+
 export {
   getActiveSpan,
   getRootSpan,
