+++ packages/browser/src/index.bundle.tracing.replay.ts
@@ -4,6 +4,8 @@ registerSpanErrorInstrumentation();
 
 export * from './index.bundle.base';
 
+export * from './metrics';
+
 export {
   getActiveSpan,
   getRootSpan,
