+++ packages/browser/src/index.bundle.tracing.ts
@@ -5,6 +5,8 @@ registerSpanErrorInstrumentation();
 
 export * from './index.bundle.base';
 
+export * from './metrics';
+
 export {
   getActiveSpan,
   getRootSpan,
