+++ packages/browser/src/index.bundle.ts
@@ -1,6 +1,7 @@
 import {
   browserTracingIntegrationShim,
   feedbackIntegrationShim,
+  metricsShim,
   replayIntegrationShim,
 } from '@sentry-internal/integration-shims';
 
@@ -11,4 +12,5 @@ export {
   feedbackIntegrationShim as feedbackAsyncIntegration,
   feedbackIntegrationShim as feedbackIntegration,
   replayIntegrationShim as replayIntegration,
+  metricsShim as metrics,
 };
