+++ packages/integration-shims/src/index.ts
@@ -1,3 +1,4 @@
 export { feedbackIntegrationShim } from './Feedback';
 export { replayIntegrationShim } from './Replay';
 export { browserTracingIntegrationShim } from './BrowserTracing';
+export { metricsShim } from './metrics';
