+++ packages/integration-shims/src/metrics.ts
@@ -1,0 +1,10 @@
+import type { Metrics } from '@sentry/types';
+
+export const metricsShim: Metrics = {
+  /* eslint-disable @typescript-eslint/no-empty-function */
+  increment: () => {},
+  distribution: () => {},
+  set: () => {},
+  gauge: () => {},
+  /* eslint-enable @typescript-eslint/no-empty-function */
+};
